# crates/crucible-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crucible web API",
    "description": "The HTTP API that `cru web` serves. This document is generated from the axum router; do not edit it by hand.",
    "version": "0.30.0"
  },
  "paths": {
...
    "/api/kilns": {
      "get": {
        "summary": "`GET /api/kilns` — every kiln a client may address.",
        "operationId": "list_kilns",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KilnListResponse"
                }
              }
            }
          },
          "502": {
            "description": "The daemon could not list the kilns, or answered a shape this route cannot read"
          }
        }
      }
    },
...
    "/api/plugins/command": {
      "post": {
        "summary": "`POST /api/plugins/command` — invoke a plugin command by name.",
        "description": "Not under `/{name}` because a command's name is already globally unique —\nthe daemon refuses a second plugin claiming one — so routing by plugin would\nask the caller for something it does not need to know. The result is passed\nthrough verbatim, like publications and options: what a command returns is\nthe plugin's vocabulary, and a shape this layer validated would be a shape\nonly today's plugins could send.",
        "operationId": "run_command",
        "parameters": [
          {
            "name": "x-crucible-plugin",
            "in": "header",
            "description": "Who is asking: `app`, or the plugin being drawn for. A request without it is refused.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginRunCommandResponse"
                }
              }
            }
          },
          "403": {
            "description": "No caller identity was sent, or the command belongs to another plugin"
          },
          "422": {
            "description": "`name` names no command"
          },
          "502": {
            "description": "The daemon could not run the command, or answered a shape this route cannot read"
          }
        }
      }
    },
    "/api/plugins/commands": {
      "get": {
        "summary": "`GET /api/plugins/commands` — the executable primitives plugins declared.",
        "description": "Each entry carries `plugin`, `name`, `description`, `hint`, `parameters`\nand `effect`. The names are read, so the browser is told what a row holds;\n`parameters` is passed through verbatim, so a caller offering a command as\na button — with a dialog built from the declared parameters — needs no\nchange on this side when a plugin ships a new one.\n\n`parameters` crosses as opaque JSON today. It comes from the same\n`ToolDefinition` a tool uses, so shaping it like `signature.rs`'s JSON\nSchema output is what would let a dialog be generated rather than\nhand-read — see `docs/Meta/Analysis/The Plugin Contract.md`.",
        "operationId": "list_commands",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginCommandsResponse"
                }
              }
            }
          },
          "502": {
            "description": "The daemon could not list the commands, or answered a shape this route cannot read"
          }
        }
      }
    },
...
    "/api/session": {
      "post": {
        "summary": "Map a `session.create` daemon error to an HTTP status. An `INVALID_PARAMS`\nerror (JSON-RPC code `-32602` — e.g. an unknown ACP profile or an\nunparseable provider override, both now resolved daemon-side) is a client\nerror (422), preserving the pre-consolidation behavior where the web\nvalidated the profile itself. Anything else is a daemon/transport failure\n(502).",
        "operationId": "create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRow"
                }
              }
            }
          },
          "422": {
            "description": "The request named an endpoint, an agent type or a card the server refuses"
          },
          "502": {
            "description": "The daemon could not create the session"
          }
        }
      }
    },
...
    "/api/session/{id}": {
      "get": {
        "operationId": "get_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The session to read",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRow"
                }
              }
            }
          },
          "502": {
            "description": "The daemon could not read the session"
          }
        }
      },
      "delete": {
        "operationId": "delete_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The session to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "No session of that id"
          },
          "502": {
            "description": "The daemon could not delete the session"
          }
        }
      }
    },
...
    "/api/session/{id}/command": {
      "post": {
        "summary": "Run one slash command in a session.",
        "operationId": "execute_command",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The session to run the command in",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "502": {
            "description": "The daemon could not serve the command"
          }
        }
      }
    },
...
    "/api/session/{id}/models": {
      "get": {
        "operationId": "list_models",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The session whose models to list",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            }
          },
          "502": {
            "description": "The daemon could not list the models"
          }
        }
      }
    },
...
    "/api/session/{id}/title": {
      "put": {
        "operationId": "set_session_title",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The session to rename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTitleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "502": {
            "description": "The daemon could not set the title"
          }
        }
      }
    },
...
    "/api/sessions/search": {
      "get": {
        "summary": "`GET /api/sessions/search?q=…&kiln=…&kiln=…&limit=…`",
        "description": "`kiln` repeats. Search scope is kiln-set *overlap*, so the caller states\nevery kiln it is cleared for rather than one member standing in for the\nrest — one member matches only the sessions sharing that one. Parsed from\nthe raw pairs because `serde_urlencoded`, which `Query` uses, cannot\ndeserialize a repeated key into a sequence.",
        "operationId": "search_sessions",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "The substring to match, case-insensitive",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kiln",
            "in": "query",
            "description": "The caller's whole kiln set, one `kiln` key per member",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "How many matches to return. The default is 20",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "No `q`, or every `kiln` named an unusable name"
          },
          "502": {
            "description": "The daemon could not run the search"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandRequest": {
        "type": "object",
        "description": "One plugin command invocation.",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "args": {}
        }
      },
      "CommandResponse": {
        "type": "object",
        "description": "What one slash command produced.\n\nA command the server does not know, and a command used wrongly, both come\nback here with `type` reading `error`: the composer prints the text either\nway, and neither is a transport failure.",
        "required": [
          "result",
          "type"
        ],
        "properties": {
          "result": {
            "type": "string",
            "description": "The text the composer prints."
          },
          "type": {
            "type": "string",
            "description": "`success` or `error`."
          }
        }
      },
...
      "CreateSessionRequest": {
        "type": "object",
        "properties": {
          "session_type": {
            "type": "string"
          },
          "kilns": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The session's kiln set by registry NAME — flat, no member privileged.\nEmpty or omitted is a literal empty set, NOT a request for a default:\nit creates a session with no corpus attached. The daemon stopped\nsubstituting its data root here because that root is the parent of the\nsessions store, so \"default\" quietly put every transcript in scope."
          },
          "workspace": {
            "type": [
              "string",
              "null"
            ]
          },
          "provider": {
            "type": [
              "string",
              "null"
            ],
            "description": "LLM provider (e.g., \"ollama\", \"openai\", \"anthropic\")"
          },
          "model": {
            "type": [
              "string",
              "null"
            ],
            "description": "Model name (e.g., \"llama3.2\", \"gpt-4o\", \"claude-3-5-sonnet\")"
          },
          "endpoint": {
            "type": [
              "string",
              "null"
            ],
            "description": "Custom endpoint URL (optional, for self-hosted models)"
          },
          "agent_type": {
            "type": [
              "string",
              "null"
            ],
            "description": "\"internal\" (default) or \"acp\""
          },
          "agent_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "ACP agent profile name (e.g. \"claude\", \"opencode\"); required when agent_type == \"acp\""
          },
          "agent_card": {
            "type": [
              "string",
              "null"
            ],
            "description": "Internal-agent card name; never resolved in the web layer."
          },
          "isolation": {
            "type": [
              "object",
              "null"
            ],
            "description": "Isolation override: absent → resolve normally; `false` → no container\neven if the project has one; `true`, a profile name or an environment\nobject → override. Forwarded to the daemon untouched — the vocabulary\nbelongs to the plugin that resolves it, and an unknown profile comes\nback as `-32602`, which `daemon_err` turns into a 422."
          }
        }
      },
      "DeleteResponse": {
        "type": "object",
        "description": "Response for session deletion.",
        "required": [
          "deleted"
        ],
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        }
      },
...
      "ExecuteCommandRequest": {
        "type": "object",
        "required": [
          "command"
        ],
        "properties": {
          "command": {
            "type": "string",
            "description": "The command line, with or without its leading slash."
          }
        }
      },
...
      "KilnListResponse": {
        "type": "object",
        "description": "What `GET /api/kilns` answers.",
        "required": [
          "kilns"
        ],
        "properties": {
          "kilns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KilnRow"
            }
          }
        }
      },
...
      "ModelsResponse": {
        "type": "object",
        "description": "Response for model listings — the session-scoped `list_models` and the\nsession-less `list_all_models` return the same `{ models: [...] }` shape.",
        "required": [
          "models"
        ],
        "properties": {
          "models": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "OkResponse": {
        "type": "object",
        "description": "Standard acknowledgment response for successful mutations.",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "PluginCommandsResponse": {
        "type": "object",
        "description": "What `GET /api/plugins/commands` answers.",
        "required": [
          "commands"
        ],
        "properties": {
          "commands": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PluginCommandRow"
            }
          }
        }
      },
...
      "PluginRunCommandResponse": {
        "type": "object",
        "description": "What `POST /api/plugins/command` answers.\n\nThe browser's hand-written caller types this as a bare `unknown`, so a\nreader had to know to look under `result` without anything saying so. The\nenvelope is the daemon's (`server/plugins.rs`'s `handle_plugin_run_command`),\nand naming it costs the plugin nothing: `result` stays opaque.",
        "required": [
          "name",
          "result"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The command that ran, echoed."
          },
          "result": {
            "description": "Whatever the command's Lua `fn` returned.\n\nOpaque, like publications and options: what a command returns is the\nplugin's vocabulary, and a shape validated here would be one only\ntoday's plugins could send."
          }
        }
      },
...
      "SessionRow": {
        "type": "object",
        "description": "One session, as every session route answers it.\n\nThree daemon methods build three different objects for one entity.\n`session.create` sends no `started_at` and no `title`. `session.get` nests\nthe model under `agent` and sends no `event_count`, `last_activity` or\n`archived`. `session.list` sends the model flat and sends all three. This\ndeclares the union once, so a client reads one type instead of reconciling\nthree. The divergence is a daemon bug (`server/session/list.rs:155` against\n`:302`); this route does not fix it.\n\nA field that one shape omits and another sends as `null` carries\n`Option<Option<T>>`. The outer level is \"the daemon wrote no key\", the\ninner one is \"the key is null\". Both spellings then reach the browser as\nthe daemon wrote them, which is what keeps this projection lossless.",
        "required": [
          "session_id",
          "type",
          "kilns",
          "state"
        ],
        "properties": {
          "session_id": {
            "type": "string",
            "description": "The session id. The wire name is `session_id`, never `id`."
          },
          "type": {
            "type": "string",
            "description": "The session type prefix, such as `chat`."
          },
          "kilns": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Every kiln the session can query, by registry name."
          },
          "workspace": {
            "type": [
              "string",
              "null"
            ],
            "description": "The session's working directory. `null` is a session with no workspace."
          },
          "state": {
            "type": "string",
            "description": "The session state, such as `active` or `paused`."
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "When the session started. `session.create` does not send it."
          },
          "title": {
            "type": [
              "string",
              "null"
            ],
            "description": "The session title. `session.create` does not send it."
          },
          "agent_model": {
            "type": [
              "string",
              "null"
            ],
            "description": "The resolved model. `session.get` does not send it; read `agent.model`\nthere."
          },
          "last_activity": {
            "type": [
              "string",
              "null"
            ],
            "description": "The last event's time. Only `session.list` sends it."
          },
          "event_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "How many events the transcript holds. Only `session.list` sends it.",
            "minimum": 0
          },
          "archived": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Whether the session is archived. Only `session.list` sends it."
          },
          "parent_session_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The parent of a delegated session. `session.create` does not send it."
          },
          "continued_from": {
            "type": [
              "string",
              "null"
            ],
            "description": "The session this one continues. Only `session.get` sends it."
          },
          "agent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SessionAgentRow",
                "description": "The session's agent record. Only `session.get` sends it."
              }
            ]
          },
          "recording_mode": {
            "type": [
              "string",
              "null"
            ],
            "description": "How the session records its transcript. `session.get` sends it only\nwhen the session has a recording mode."
          }
        }
      },
...
      "SessionSearchMatch": {
        "type": "object",
        "description": "One transcript line that matched a session search.\n\nNot a session: the daemon answers the line it matched on, so a caller that\nwants the session reads `session_id` and asks for it.",
        "required": [
          "session_id",
          "line",
          "context"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "line": {
            "type": "integer",
            "format": "int64",
            "description": "The 1-based line of the transcript. `0` marks a title match on a\nsession whose transcript has not reached disk yet.",
            "minimum": 0
          },
          "context": {
            "type": "string",
            "description": "The matched line, truncated to 100 characters."
          }
        }
      },
      "SessionSearchResponse": {
        "type": "object",
        "description": "What `GET /api/sessions/search` answers.",
        "required": [
          "matches",
          "total"
        ],
        "properties": {
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionSearchMatch"
            }
          },
          "total": {
            "type": "integer",
            "description": "How many matches the reply carries.",
            "minimum": 0
          },
          "note": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why the search looked at nothing, when it looked at nothing.\n\nThe daemon writes it for a search with no kiln scope\n(`server/session/list.rs:206`), and only then. An unscoped search is\nthe one case where an empty result is not a statement about the\ncorpus, so the sentence has to reach the caller."
          }
        }
      },
...
      "SetTitleRequest": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string"
          }
        }
      },
...
    }
  }
}